20.03.2016 H

+++ NaftanRailway.WebUI/ViewModels/PagingInfo.cs
@@ -5,7 +5,7 @@ namespace NaftanRailway.WebUI.ViewModels {
     /// View model. For passing information about paging between model and view throught action
     /// </summary>
     public class PagingInfo {
-        public int TotalItems { get; set; }
+        public long TotalItems { get; set; }
         public int ItemsPerPage { get; set; }
         public int CurrentPage { get; set; }
